copied stitcher impl from simple-tp-stitcher app
